requirements.txt
numpy==1.19.5
scipy==1.7.1
pillow==8.3.2
scikit-image==0.18.3
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.9.0+cpu
torchvision==0.10.0+cpu
